# Milestone 4/backend/swagger.yaml
openapi: 3.1.0
info:
  title: EduAssist API
  description: AI-Powered Teaching Assistant Platform - Knowledge Assistant Feature
  version: 1.0.0
paths:
...
  /assessments/assessments/:
    post:
      tags:
        - Assessments
        - Assessment Generator
      summary: Create Assessment
      description: |-
        Create AI-generated assessment

        Steps:
        1. Create assessment record (GENERATING status)
        2. Generate questions in background
        3. Return assessment ID immediately

        Client should poll /assessments/{id} to check status
      operationId: create_assessment_assessments_assessments__post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Assessments
        - Assessment Generator
      summary: List Assessments
      description: List assessments
      operationId: list_assessments_assessments_assessments__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: course_id
          in: query
          required: false
          schema:
            type: integer
            title: Course Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /assessments/assessments/{assessment_id}:
    get:
      tags:
        - Assessments
        - Assessment Generator
      summary: Get Assessment
      description: Get assessment by ID (with answers - for instructors)
      operationId: get_assessment_assessments_assessments__assessment_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: assessment_id
          in: path
          required: true
          schema:
            type: integer
            title: Assessment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Assessments
        - Assessment Generator
      summary: Delete Assessment
      description: Delete assessment
      operationId: delete_assessment_assessments_assessments__assessment_id__delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: assessment_id
          in: path
          required: true
          schema:
            type: integer
            title: Assessment Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /assessments/assessments/{assessment_id}/preview:
    get:
      tags:
        - Assessments
        - Assessment Generator
      summary: Preview Assessment
      description: |-
        Get assessment preview (no answers - for students)

        Used by students to take the assessment
      operationId: preview_assessment_assessments_assessments__assessment_id__preview_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: assessment_id
          in: path
          required: true
          schema:
            type: integer
            title: Assessment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentPreview'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /assessments/assessments/{assessment_id}/attempts:
    post:
      tags:
        - Assessments
        - Assessment Generator
      summary: Start Attempt
      description: Start an assessment attempt
      operationId: start_attempt_assessments_assessments__assessment_id__attempts_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: assessment_id
          in: path
          required: true
          schema:
            type: integer
            title: Assessment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentAttemptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /assessments/assessments/attempts/{attempt_id}/submit:
    post:
      tags:
        - Assessments
        - Assessment Generator
      summary: Submit Attempt
      description: Submit assessment answers
      operationId: submit_attempt_assessments_assessments_attempts__attempt_id__submit_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: attempt_id
          in: path
          required: true
          schema:
            type: integer
            title: Attempt Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentSubmit'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentAttemptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /assessments/assessments/my-attempts:
    get:
      tags:
        - Assessments
        - Assessment Generator
      summary: Get My Attempts
      description: Get current user's assessment attempts
      operationId: get_my_attempts_assessments_assessments_my_attempts_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: assessment_id
          in: query
          required: false
          schema:
            type: integer
            title: Assessment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssessmentAttemptResponse'
                title: Response Get My Attempts Assessments Assessments My Attempts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnswerSubmission:
      properties:
        question_id:
          type: integer
          title: Question Id
        answer:
          title: Answer
      type: object
      required:
        - question_id
        - answer
      title: AnswerSubmission
      description: Student's answer to a question
    AssessmentAttemptResponse:
      properties:
        id:
          type: integer
          title: Id
        assessment_id:
          type: integer
          title: Assessment Id
        student_id:
          type: integer
          title: Student Id
        score:
          anyOf:
            - type: number
            - type: 'null'
          title: Score
        max_score:
          type: integer
          title: Max Score
        percentage:
          anyOf:
            - type: number
            - type: 'null'
          title: Percentage
        started_at:
          type: string
          format: date-time
          title: Started At
        submitted_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Submitted At
        time_taken_minutes:
          anyOf:
            - type: integer
            - type: 'null'
          title: Time Taken Minutes
      type: object
      required:
        - id
        - assessment_id
        - student_id
        - max_score
        - started_at
      title: AssessmentAttemptResponse
      description: Assessment attempt result
    AssessmentCreate:
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        course_id:
          type: integer
          title: Course Id
        custom_prompt:
          type: string
          minLength: 10
          title: Custom Prompt
          description: Instructions for question generation
        difficulty_level:
          $ref: '#/components/schemas/DifficultyLevel'
          default: MEDIUM
        question_types:
          items:
            $ref: '#/components/schemas/QuestionType'
          type: array
          minItems: 1
          title: Question Types
        total_questions:
          type: integer
          maximum: 100
          minimum: 1
          title: Total Questions
        reference_document_ids:
          items:
            type: integer
          type: array
          title: Reference Document Ids
          description: Document IDs to use as reference
        duration_minutes:
          anyOf:
            - type: integer
              maximum: 600
              minimum: 5
            - type: 'null'
          title: Duration Minutes
      type: object
      required:
        - title
        - course_id
        - custom_prompt
        - question_types
        - total_questions
      title: AssessmentCreate
      description: Request to create assessment
    AssessmentListResponse:
      properties:
        assessments:
          items:
            $ref: '#/components/schemas/AssessmentResponse'
          type: array
          title: Assessments
        total:
          type: integer
          title: Total
      type: object
      required:
        - assessments
        - total
      title: AssessmentListResponse
      description: List of assessments
    AssessmentPreview:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        total_questions:
          type: integer
          title: Total Questions
        total_marks:
          type: integer
          title: Total Marks
        duration_minutes:
          anyOf:
            - type: integer
            - type: 'null'
          title: Duration Minutes
        questions:
          items:
            $ref: '#/components/schemas/QuestionPreview'
          type: array
          title: Questions
          default: []
      type: object
      required:
        - id
        - title
        - total_questions
        - total_marks
      title: AssessmentPreview
      description: Assessment for students (no answers)
    AssessmentResponse:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        custom_prompt:
          type: string
          title: Custom Prompt
        difficulty_level:
          $ref: '#/components/schemas/DifficultyLevel'
        question_types:
          items:
            type: string
          type: array
          title: Question Types
        total_questions:
          type: integer
          title: Total Questions
        status:
          $ref: '#/components/schemas/AssessmentStatus'
        total_marks:
          anyOf:
            - type: integer
            - type: 'null'
          title: Total Marks
        duration_minutes:
          anyOf:
            - type: integer
            - type: 'null'
          title: Duration Minutes
        course_id:
          type: integer
          title: Course Id
        created_by:
          type: integer
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        questions:
          items:
            $ref: '#/components/schemas/QuestionResponse'
          type: array
          title: Questions
          default: []
      type: object
      required:
        - id
        - title
        - custom_prompt
        - difficulty_level
        - question_types
        - total_questions
        - status
        - course_id
        - created_by
        - created_at
      title: AssessmentResponse
      description: Assessment response
    AssessmentStatus:
      type: string
      enum:
        - GENERATING
        - COMPLETED
        - FAILED
      title: AssessmentStatus
    AssessmentSubmit:
      properties:
        answers:
          items:
            $ref: '#/components/schemas/AnswerSubmission'
          type: array
          title: Answers
      type: object
      required:
        - answers
      title: AssessmentSubmit
      description: Submit assessment answers
...
    DifficultyLevel:
      type: string
      enum:
        - EASY
        - MEDIUM
        - HARD
      title: DifficultyLevel
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QuestionPreview:
      properties:
        id:
          type: integer
          title: Id
        question_type:
          $ref: '#/components/schemas/QuestionType'
        question_text:
          type: string
          title: Question Text
        options:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Options
        marks:
          type: integer
          title: Marks
        order:
          type: integer
          title: Order
      type: object
      required:
        - id
        - question_type
        - question_text
        - marks
        - order
      title: QuestionPreview
      description: Question without correct answer (for students)
    QuestionResponse:
      properties:
        id:
          type: integer
          title: Id
        assessment_id:
          type: integer
          title: Assessment Id
        question_type:
          $ref: '#/components/schemas/QuestionType'
        question_text:
          type: string
          title: Question Text
        options:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Options
        correct_answer:
          title: Correct Answer
        explanation:
          anyOf:
            - type: string
            - type: 'null'
          title: Explanation
        marks:
          type: integer
          title: Marks
        difficulty:
          $ref: '#/components/schemas/DifficultyLevel'
        order:
          type: integer
          title: Order
      type: object
      required:
        - id
        - assessment_id
        - question_type
        - question_text
        - correct_answer
        - marks
        - difficulty
        - order
      title: QuestionResponse
      description: Question response
    QuestionType:
      type: string
      enum:
        - MCQ
        - MSQ
        - NAT
        - SHORT_ANSWER
        - TRUE_FALSE
      title: QuestionType
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token